Collapse QueryStore abstraction into QueryManager.
QueryStore information belongs in QueryInfo objects, and getting rid of
the QueryStore abstraction is a first step in that direction.
